Support vars that is a JSON dump of FEEL context
We may want the GraphQL mutation to be able to see the entire FEEL
context (all job info and process instance variables) for
situations such as saving entire context state to database, or
if the mutation handler on the GraphQL server can do more work.

This commit allows the BPMN modeler to define a `graphql_var_`
prefixed header with the value `=$` so this job worker
will JSON stringify the entire FEEL context into a GraphQL
query variable.