GraphQL query data is flattened to Zeebe Variables
Before, only top level GraphQL data keys would be created
as Zeebe Process Instance Variables, and their nested data
would be as JSON string. This would make it harder to
deal with nested data.

This change flattens the GraphQL returned data so that
all nested data will be created as top level Instance Variables,
but named accordingly to the nested structure.

We also change the semantics of the `graphql_data_key`, such that
if set, then a top level Process Instance Variable is set with
the raw data of the GraphQL query result (of the key's value);
otherwise it is unset.